requirements.txt
pandas
streamlit-aggrid
altair
plotly
SQLAlchemy
